+++ crates/core/src/crypto.rs
@@ -2,16 +2,23 @@
 // SPDX-License-Identifier: Apache-2.0
 
 //! Cryptographic types for signing messages.
-use anyhow::Result;
-use bip32::Mnemonic;
+use anyhow::{Result, bail};
+use bip39::Mnemonic;
 use blake2::{Blake2s, Digest, digest, digest::typenum::ToInt};
-use ed25519_dalek::{SecretKey, Signer, Verifier};
+use ed25519_dalek::{Signer, Verifier};
 use rand::{CryptoRng, RngCore, SeedableRng, rngs::StdRng};
 use serde::{Deserialize, Serialize};
 use std::fmt;
+use zeroize::Zeroizing;
+
+const ENTROPY_LEN: usize = 16;
+type Entropy = [u8; ENTROPY_LEN];
 
 /// A key for signing messages.
-pub struct SigningKey(ed25519_dalek::SigningKey);
+pub struct SigningKey {
+    key: ed25519_dalek::SigningKey,
+    entropy: Zeroizing<Entropy>,
+}
 
 /// The hasher used for signatures.
 type SigHasher = Blake2s<digest::consts::U32>;
@@ -29,10 +36,14 @@ impl SigningKey {
     where
         S: AsRef<str>,
     {
-        let mnemonic = Mnemonic::new(phrase, Default::default())?;
-        Ok(Self(ed25519_dalek::SigningKey::from_bytes(
-            mnemonic.entropy(),
-        )))
+        let mnemonic = Mnemonic::from_phrase(phrase.as_ref(), Default::default())?;
+        if mnemonic.entropy().len() != ENTROPY_LEN {
+            bail!("Invalid passphrase length");
+        }
+
+        let mut entropy = Entropy::default();
+        entropy.copy_from_slice(mnemonic.entropy());
+        Ok(Self::from_entropy(entropy))
     }
 
     /// Sign a message.
@@ -42,24 +53,35 @@ impl SigningKey {
     {
         let mut hasher = SigHasher::new();
         bincode::serialize_into(&mut hasher, msg).expect("should serialize to hasher");
-        Signature(self.0.sign(&hasher.finalize()))
+        Signature(self.key.sign(&hasher.finalize()))
     }
 
     /// Get the secret key phrase.
     pub fn phrase(&self) -> String {
-        let mnemonic = Mnemonic::from_entropy(*self.0.as_bytes(), Default::default());
-        mnemonic.phrase().to_string()
+        // This should never fail as we control the entropy size.
+        Mnemonic::from_entropy(self.entropy.as_ref(), Default::default())
+            .unwrap()
+            .phrase()
+            .to_string()
     }
 
     /// Get the signature verifying key.
     pub fn verifying_key(&self) -> VerifyingKey {
-        VerifyingKey(self.0.verifying_key())
+        VerifyingKey(self.key.verifying_key())
     }
 
     fn from_crypto_rng<R: RngCore + CryptoRng>(rng: &mut R) -> Self {
-        let mut key = SecretKey::default();
-        rng.fill_bytes(&mut key);
-        Self(ed25519_dalek::SigningKey::from_bytes(&key))
+        let mut entropy = Entropy::default();
+        rng.fill_bytes(&mut entropy);
+        Self::from_entropy(entropy)
+    }
+
+    fn from_entropy(entropy: Entropy) -> Self {
+        // Hash 128 bits entropy to 256 bits SigningKey.
+        let key_hash = SigHasher::digest(entropy);
+        let key = ed25519_dalek::SigningKey::from_bytes(&key_hash.into());
+        let entropy = Zeroizing::new(entropy);
+        Self { key, entropy }
     }
 }
 
@@ -68,7 +90,7 @@ impl fmt::Debug for SigningKey {
         write!(
             f,
             "SigningKey({})",
-            bs58::encode(self.0.as_bytes()).into_string()
+            bs58::encode(self.key.as_bytes()).into_string()
         )
     }
 }
@@ -156,7 +178,7 @@ mod tests {
     fn keypair_phrase() {
         let sk = SigningKey::default();
         let from_phrase = SigningKey::from_phrase(sk.phrase()).unwrap();
-        assert_eq!(sk.0, from_phrase.0);
+        assert_eq!(sk.key, from_phrase.key);
     }
 
     #[test]